PoC for using stroke-dasharray for Area w/ line
The current Area component renders the path twice if
line=true is used. Once for the area and again for
the line on the top of the area.

This change removes the second path and replaces w/
a stroke-dasharray implementation. This reduces the
DOM size significantly since we're no longer rendering
2 path elements and instead are drawing the line using
CSS stroke-dasharray. This is implemented by measuring
the total area's path length and then subtracting the
left, right, and bottom 'legs' of the path.

** Do not merge **. This is juyst a PoC to show how to
implement such a thing and more care needs to be made
to get it working with existing interfaces (e.g. stroke
styles for the line.)