Column
Broken ones have been commented out in md file.
Multiple of same error related to LayerContext

Check data export on stacked versions, not sure how it should be done.

stacked-percent, grouped-and-stacked, grouped-stacked-or-both-transition,  tooltip-and-click-handlers-for-individual-stack-grouped-bar, broken with "TypeError: The first argument of calcUniques() must be an array. You passed in a object. If you got this error using the Chart component, consider passing a flat array to the flatData prop" - Ithink AI bite me here but could get it working.